Revert "Move MapSO parsers into their own file"
This reverts commit e81b50f2e93672dacaecd14457b19b4e22ccf62f.
